requirements.txt
pandas
numpy
spacy~=2.1.9
cytoolz
bs4
aiohttp
matplotlib
PyYAML
scipy